moved get_video_shape from extraction_tools to numpyextractors where it is used
